Split middleware into discovery and action
By separating discovery from resolving and invoking, it is possible to
apply routing earlier in the middleware process to handle 404s and 405s
as quickly as possible, before doing any additional work in other
middleware.

This provides a clearer separation of concerns and allows more complex
use cases.
