+++ src/FastRouteDiscovery.php
@@ -1,0 +1,68 @@
+<?php
+
+namespace Middlewares;
+
+use Middlewares\Utils\Factory;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Message\ResponseInterface;
+use Interop\Http\ServerMiddleware\MiddlewareInterface;
+use Interop\Http\ServerMiddleware\DelegateInterface;
+use FastRoute\Dispatcher;
+
+class FastRouteDiscovery implements MiddlewareInterface
+{
+    /**
+     * @var Dispatcher FastRoute dispatcher
+     */
+    private $router;
+
+    /**
+     * Set the Dispatcher instance.
+     *
+     * @param Dispatcher $router
+     */
+    public function __construct(Dispatcher $router)
+    {
+        $this->router = $router;
+    }
+
+    /**
+     * Process a server request and return a response.
+     *
+     * @param ServerRequestInterface $request
+     * @param DelegateInterface      $delegate
+     *
+     * @return ResponseInterface
+     */
+    public function process(ServerRequestInterface $request, DelegateInterface $delegate)
+    {
+        $route = $this->router->dispatch($request->getMethod(), $request->getUri()->getPath());
+
+        if ($route[0] === Dispatcher::NOT_FOUND) {
+            return Factory::createResponse(404);
+        }
+
+        if ($route[0] === Dispatcher::METHOD_NOT_ALLOWED) {
+            return Factory::createResponse(405)->withHeader('Allow', implode(', ', $route[1]));
+        }
+
+        foreach ($route[2] as $name => $value) {
+            $request = $request->withAttribute($name, $value);
+        }
+
+        $request = $this->setController($request, $route[1]);
+
+        return $delegate->process($request);
+    }
+
+    /**
+     * @param ServerRequestInterface $request
+     * @param callable|string|array  $controller
+     *
+     * @return ServerRequestInterface
+     */
+    protected function setController(ServerRequestInterface $request, $controller)
+    {
+        return $request->withAttribute('controller', $controller);
+    }
+}
